requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pyserial==3.6
thread6==0.2.0